add new arguments
Add a rank_cols argument to load_csv so that the user can specify which columns of the csv correspond to rankings. Helpful, for example, with the 2013 Minneapolis data file on MGGG/chicago which has extraneous columns.

Also removed quota parameter from Election.get_threshold method since quota should be an attribute of the Election.