address James comments
Thanks, @jamesturk ! I added .idea to .gitignore, and gave the `rank_cols` a default argument.

I think it's worth discussing whether or not you actually want the `rank_cols` parameter to have a default argument. My argument for not including a default is that we don't want to force preprocessing of the csv file onto the user, and we want the user to think intentionally about loading in their data. Happy to hear other ideas about why we might want a default.